Autoloot also
