+++ ClassicAssist/UI/ViewModels/Agents/DraggableTreeViewViewModel.cs
@@ -1,0 +1,19 @@
+﻿#region License
+// Copyright (C) 2023 Reetus
+// 
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+// 
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY
+#endregion
+
+namespace ClassicAssist.UI.ViewModels.Agents
+{
+    public class DraggableTreeViewViewModel
+    {
+        
+    }
+}
