+++ ClassicAssist/UI/ViewModels/MacrosTabViewModel.cs
@@ -278,14 +278,15 @@ namespace ClassicAssist.UI.ViewModels
             }
 
             foreach ( IDraggableGroup draggableGroup in Draggables.Where( i => i is IDraggableGroup )
-                         .OrderBy( GetDraggableIndex ).Cast<IDraggableGroup>() )
+                         .OrderBy( e => DraggableTreeViewHelpers.GetIndex( e, Draggables ) ).Cast<IDraggableGroup>() )
             {
                 JObject entry = new JObject { { "Name", draggableGroup.Name } };
 
                 groupArray.Add( entry );
             }
 
-            IEnumerable<MacroEntry> globalMacros = Items.Where( e => e.Global ).OrderBy( GetDraggableIndex ).ToList();
+            IEnumerable<MacroEntry> globalMacros = Items.Where( e => e.Global )
+                .OrderBy( e => DraggableTreeViewHelpers.GetIndex( e, Draggables ) ).ToList();
 
             if ( globalMacros.Any() )
             {
@@ -296,7 +297,8 @@ namespace ClassicAssist.UI.ViewModels
                     globalJson );
             }
 
-            foreach ( MacroEntry macroEntry in Items.Where( e => !e.Global ).OrderBy( GetDraggableIndex ).ToList() )
+            foreach ( MacroEntry macroEntry in Items.Where( e => !e.Global )
+                         .OrderBy( e => DraggableTreeViewHelpers.GetIndex( e, Draggables ) ).ToList() )
             {
                 macroArray.Add( macroEntry.ToJObject() );
             }
@@ -437,32 +439,6 @@ namespace ClassicAssist.UI.ViewModels
             }
 
             SelectedItem = Items.LastOrDefault();
-        }
-
-        private int GetDraggableIndex( IDraggable draggable )
-        {
-            ObservableCollection<IDraggable> parent = GetParent( draggable, Draggables );
-
-            return parent?.IndexOf( draggable ) ?? Draggables.IndexOf( draggable );
-        }
-
-        private static IEnumerable<IDraggableGroup> GetGroups( IEnumerable<IDraggable> collection )
-        {
-            return collection.Where( i => i is IDraggableGroup ).Cast<IDraggableGroup>();
-        }
-
-        private static ObservableCollection<IDraggable> GetParent( IDraggable draggable,
-            ObservableCollection<IDraggable> parent )
-        {
-            if ( parent.Contains( draggable ) )
-            {
-                return parent;
-            }
-
-            IEnumerable<IDraggableGroup> groups = GetGroups( parent );
-
-            return groups.Select( draggableGroup => GetParent( draggable, draggableGroup.Children ) )
-                .FirstOrDefault( childParent => childParent != null );
         }
 
         private void NewPublicMacro( Metadata metadata )
